Completed and tested P0 implementation
* basic math with user input builds and works appropriately. WIP on adding tests and better output formatting.

* Moving cost calculation to a separate function, setting doNotSign

Moved the actual calculation of the total cost based on the number of
people to a separate function, and wrote a unit test for this function.

Also set `doNotSign` for `craneLib.buildPackage` to avoid Nix build
errors until I decide if I want to fix the signing issue.

* Using floating point value for price of potatoes

The first iteration was hardcoded as an `i32` to be the easiest possible
intro to the Rust typesystem. This change implements explicit casting to floating
point for the price calculations because very few prices are integers.

* Converting to f32 for everything and splitting out some new functions

* Formatting commit

* Updating deny.toml with new licenses from dependencies

* Finishing P0 implementation

- Converting all usages of numbers to Decimal types to avoid various casting operations
- Adding new `extract_input()` function to get the input from user at
  command line
- Adding unit tests for new function
- Modifying unit tests to use Decimal input and output